- Updated folder structures - Split files - Fix codes
