+++ schema/club.go
@@ -7,9 +7,3 @@ type Club struct {
 	MonthlyFee float64 `json:"monthly_fee"`
 	RegNo      string  `json:"reg_no"`
 }
-
-type Response struct {
-	Message string      `json:"message"`
-	Data    interface{} `json:"data,omitempty"`
-	Error   string      `json:"error,omitempty"`
-}
