+++ schema/response.go
@@ -1,0 +1,7 @@
+package schemas
+
+type Response struct {
+	Message string      `json:"message"`
+	Data    interface{} `json:"data,omitempty"`
+	Error   string      `json:"error,omitempty"`
+}
